refactor: move `JobState` to `job_state.rs`
Also add two new methods:

- `JobState::new` to create new `JobState`
- `JobState::run_to_finish` is extracted from old [`doit`][1].

[1]: https://github.com/rust-lang/cargo/blob/7ddcf0fe3348b7e0f4ad4a730eab60a20638ef28/src/cargo/core/compiler/job_queue.rs#L1122-L1168
